get num of frames faster and signal datatype from float16 to uint16

+++ src/cardiomea/pipelines/features/nodes.py
@@ -153,10 +153,10 @@ def extract_data(file_path_full, start_frame, length, s_freq):
     if start_frame < (num_frames-1):
         if (start_frame+length) > num_frames:
             warnings.warn("The specified time frame exceeds the data length. Signals will be extracted until end-of-file.")
-        signals = (obj.get('sig')[channel_ids,start_frame:start_frame+length] * lsb).astype('float16')
+        signals = (obj.get('sig')[channel_ids,start_frame:start_frame+length] * lsb).astype('uint16')
     else:
         warnings.warn(f"The start frame exceeds the length of data. Signals will be extracted from the start of the recording until MIN({length}-th frame, end-of-file) instead.")
-        signals = (obj.get('sig')[channel_ids,0:min(length,num_frames)] * lsb).astype('float16')
+        signals = (obj.get('sig')[channel_ids,0:min(length,num_frames)] * lsb).astype('uint16')
 
     return signals, electrodes_info, gain, num_frames/s_freq, signals.shape[1]/s_freq
